[Core] Cuts generic _make method renamed to _makeDefault [Core] CacheLine collections updated to use less memory [Attack] Attack hold set fix while move

+++ Runtime/Jobs/Jobs.cs
@@ -112,7 +112,7 @@ namespace ME.BECS {
 
         public const uint CacheLineSize = JobsUtility.CacheLineSize;
         public static uint ThreadsCount => (uint)JobsUtility.ThreadIndexCount;
-        public static int ThreadIndex => JobsUtility.ThreadIndex;
+        public static uint ThreadIndex => (uint)JobsUtility.ThreadIndex;
 
         public static void Initialize() {
             CleanUp();
